refactor: split construct tests
